requirements.txt
moviepy
matplotlib
tqdm
opencv-python
Pillow
